fixed some issues

+++ app/Services/Products/CategoryService.php
@@ -14,19 +14,30 @@ class CategoryService
     {
         $categories = DB::table('categories')->where('parent_category_id', null)->orderBy('name', 'asc')->get();
 
-        $img_url = asset('uploads/category/');
-
         return DataTables::of($categories)
             ->addIndexColumn()
-            ->editColumn('photo', function ($row) use ($img_url) {
+            ->editColumn('photo', function ($row) {
 
-                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;"  src="' . $img_url . '/' . $row->photo . '">';
+                $photo = asset('images/general_default.png');
+                if ($row->photo) {
+
+                    $photo = asset('uploads/category/' . $row->photo);
+                }
+                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;"  src="' . $photo . '">';
             })
             ->addColumn('action', function ($row) {
 
                 $html = '<div class="dropdown table-dropdown">';
-                $html .= '<a href="' . route('categories.edit', [$row->id]) . '" class="action-btn c-edit" id="editCategory" title="Edit"><span class="fas fa-edit"></span></a>';
-                $html .= '<a href="' . route('categories.delete', [$row->id]) . '" class="action-btn c-delete" id="deleteCategory" title="Delete"><span class="fas fa-trash "></span></a>';
+
+                if (auth()->user()->can('product_category_edit')) {
+
+                    $html .= '<a href="' . route('categories.edit', [$row->id]) . '" class="action-btn c-edit" id="editCategory" title="Edit"><span class="fas fa-edit"></span></a>';
+                }
+
+                if (auth()->user()->can('product_category_delete')) {
+
+                    $html .= '<a href="' . route('categories.delete', [$row->id]) . '" class="action-btn c-delete" id="deleteCategory" title="Delete"><span class="fas fa-trash "></span></a>';
+                }
                 $html .= '</div>';
 
                 return $html;
@@ -67,18 +78,15 @@ class CategoryService
 
         if ($request->file('photo')) {
 
-            if ($updateCategory->photo !== 'default.png') {
+            if ($updateCategory->photo && file_exists(public_path('uploads/category/' . $updateCategory->photo))) {
 
-                if (file_exists(public_path('uploads/category/' . $updateCategory->photo))) {
-
-                    unlink(public_path('uploads/category/' . $updateCategory->photo));
-                }
+                unlink(public_path('uploads/category/' . $updateCategory->photo));
             }
 
             $categoryPhoto = $request->file('photo');
             $categoryPhotoName = uniqid() . '.' . $categoryPhoto->getClientOriginalExtension();
             Image::make($categoryPhoto)->resize(250, 250)->save('uploads/category/' . $categoryPhotoName);
-            $updateCategory->phone = $categoryPhotoName;
+            $updateCategory->photo = $categoryPhotoName;
         }
 
         $updateCategory->save();
@@ -95,12 +103,9 @@ class CategoryService
             return ['pass' => false, 'msg' => 'Category can not be deleted. One or more sub-categories is belonging under this category.'];
         }
 
-        if ($deleteCategory->photo !== 'default.png') {
+        if ($deleteCategory->photo && file_exists(public_path('uploads/category/' . $deleteCategory->photo))) {
 
-            if (file_exists(public_path('uploads/category/' . $deleteCategory->photo))) {
-
-                unlink(public_path('uploads/category/' . $deleteCategory->photo));
-            }
+            unlink(public_path('uploads/category/' . $deleteCategory->photo));
         }
 
         if (!is_null($deleteCategory)) {
