+++ app/Services/Products/SubCategoryService.php
@@ -14,22 +14,38 @@ class SubCategoryService
         $subCategories = DB::table('categories')
             ->leftJoin('categories as parentCategory', 'categories.id', 'parentCategory.id')
             ->whereNotNull('categories.parent_category_id')
-            ->select('categories.id', 'categories.name', 'categories.photo', 'categories.description', 'parentCategory.name as parent_category_name')
-            ->orderBy('id', 'desc')->get();
-
-        $imgUrl = asset('uploads/category/');
+            ->select(
+                'categories.id',
+                'categories.name',
+                'categories.photo',
+                'categories.description',
+                'parentCategory.name as parent_category_name'
+            )->orderBy('id', 'desc')->get();
 
         return DataTables::of($subCategories)
             ->addIndexColumn()
-            ->editColumn('photo', function ($row) use ($imgUrl) {
+            ->editColumn('photo', function ($row) {
 
-                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;"  src="'.$imgUrl.'/'.$row->photo.'">';
+                $photo = asset('images/general_default.png');
+                if ($row->photo) {
+
+                    $photo = asset('uploads/category/' . $row->photo);
+                }
+                return '<img loading="lazy" class="rounded img-thumbnail" style="height:30px; width:30px;"  src="' . $photo . '">';
             })
             ->addColumn('action', function ($row) {
 
                 $html = '<div class="dropdown table-dropdown">';
-                $html .= '<a href="'.route('subcategories.edit', [$row->id]).'" class="action-btn c-edit" id="editSubcategory"><span class="fas fa-edit"></span></a>';
-                $html .= '<a href="'.route('subcategories.delete', [$row->id]).'" class="action-btn c-delete" id="deleteSubcategory"><span class="fas fa-trash "></span></a>';
+
+                if (auth()->user()->can('product_category_edit')) {
+
+                    $html .= '<a href="' . route('subcategories.edit', [$row->id]) . '" class="action-btn c-edit" id="editSubcategory"><span class="fas fa-edit"></span></a>';
+                }
+
+                if (auth()->user()->can('product_category_delete')) {
+
+                    $html .= '<a href="' . route('subcategories.delete', [$row->id]) . '" class="action-btn c-delete" id="deleteSubcategory"><span class="fas fa-trash "></span></a>';
+                }
                 $html .= '</div>';
 
                 return $html;
@@ -47,10 +63,10 @@ class SubCategoryService
 
         if ($request->file('photo')) {
 
-            $categoryPhoto = $request->file('photo');
-            $categoryPhotoName = uniqid().'.'.$categoryPhoto->getClientOriginalExtension();
-            Image::make($categoryPhoto)->resize(250, 250)->save('uploads/category/'.$categoryPhotoName);
-            $addSubCategory = $categoryPhotoName;
+            $subcategoryPhoto = $request->file('photo');
+            $subcategoryPhotoName = uniqid() . '.' . $subcategoryPhoto->getClientOriginalExtension();
+            Image::make($subcategoryPhoto)->resize(250, 250)->save('uploads/category/' . $subcategoryPhotoName);
+            $addSubCategory->photo = $subcategoryPhotoName;
         }
 
         $addSubCategory->save();
@@ -67,18 +83,15 @@ class SubCategoryService
 
         if ($request->file('photo')) {
 
-            if ($updateSubcategory->photo !== 'default.png') {
+            if ($updateSubcategory->photo && file_exists(public_path('uploads/category/' . $updateSubcategory->photo))) {
 
-                if (file_exists(public_path('uploads/category/'.$updateSubcategory->photo))) {
-
-                    unlink(public_path('uploads/category/'.$updateSubcategory->photo));
-                }
+                unlink(public_path('uploads/category/' . $updateSubcategory->photo));
             }
 
-            $categoryPhoto = $request->file('photo');
-            $categoryPhotoName = uniqid().'.'.$categoryPhoto->getClientOriginalExtension();
-            Image::make($categoryPhoto)->resize(250, 250)->save('uploads/category/'.$categoryPhotoName);
-            $updateSubcategory->photo = $categoryPhotoName;
+            $subcategoryPhoto = $request->file('photo');
+            $subcategoryPhotoName = uniqid() . '.' . $subcategoryPhoto->getClientOriginalExtension();
+            Image::make($subcategoryPhoto)->resize(250, 250)->save('uploads/category/' . $subcategoryPhotoName);
+            $updateSubcategory->photo = $subcategoryPhotoName;
         }
 
         $updateSubcategory->save();
@@ -90,14 +103,11 @@ class SubCategoryService
     {
         $deleteSubcategory = $this->singleSubcategory(id: $id);
 
-        if (isset( $deleteSubcategory)) {
+        if (isset($deleteSubcategory)) {
 
-            if ($deleteSubcategory->photo !== 'default.png') {
+            if ($deleteSubcategory->photo && file_exists(public_path('uploads/category/' . $deleteSubcategory->photo))) {
 
-                if (file_exists(public_path('uploads/category/'.$deleteSubcategory->photo))) {
-
-                    unlink(public_path('uploads/category/'.$deleteSubcategory->photo));
-                }
+                unlink(public_path('uploads/category/' . $deleteSubcategory->photo));
             }
 
             $deleteSubcategory->delete();
